Threads are now managed and close on command.
Each class that creates threads will now save the Thread object so it can later interrupt it.
(ServerServer::createAndStartNewThread does this.)
(Except for HeadTCPThread which manages its own list since they are always being created/destroyed)

ServerServer::close will close the original sockets, and then interrupt any threads is created.
ServerServer::close can optionally wait until all threads have returned.

IoTPacketSection now adds a null byte to the end of strings.
Except for ZIoT, which is the header and doesn't need one.

+++ src/com/zoarial/JavaIoTAction.java
@@ -7,7 +7,7 @@ import java.util.function.Supplier;
 public class JavaIoTAction extends IoTAction {
     Supplier<String> exec;
 
-    public JavaIoTAction(String name, UUID uuid, byte level, int args, Supplier<String> exec) {
+    public JavaIoTAction(String name, UUID uuid, byte level, byte args, Supplier<String> exec) {
         super(name, uuid, level, args);
         this.exec = exec;
     }
